wode

+++ blog/adminrelate/poempost.py
@@ -15,7 +15,7 @@ import re
 __author__ = 'benywon'
 
 showlen=15
-
+preeditid=-1
 SOURCES_CHOICES = (
                   (False, '不是以前的词'),
                   (True, '是以前的词'),
@@ -23,9 +23,16 @@ SOURCES_CHOICES = (
 class PoemForm(forms.Form):
     title = forms.CharField(label=u'词牌名',max_length=100,widget=forms.TextInput(attrs={'size':30}))
     nav = forms.CharField(label=u'叙述',max_length=200,required=False,widget=forms.Textarea(attrs={'style':"height:51px;width:249px"}))
-    poem=forms.CharField(label='诗歌',widget=forms.Textarea(attrs={'style':"height:200px;width:849px"}),max_length=1200)
+    poem=forms.CharField(label=u'诗歌',widget=forms.Textarea(attrs={'style':"height:200px;width:849px"}),max_length=1200)
     isprevious = forms.ChoiceField(label=u'是否是以前',choices=SOURCES_CHOICES)
-
+class PoemFormModel(forms.ModelForm):
+    title = forms.CharField(label=u'词牌名',max_length=100,widget=forms.TextInput(attrs={'size':30}))
+    nav = forms.CharField(label=u'叙述',max_length=200,required=False,widget=forms.Textarea(attrs={'style':"height:51px;width:249px"}))
+    # poem=forms.CharField(label=u'诗歌',widget=forms.Textarea(attrs={'style':"height:200px;width:849px"}),max_length=1200)
+    isprevious = forms.ChoiceField(label=u'是否是以前',choices=SOURCES_CHOICES)
+    class Meta:
+        model=BlogPoem
+        fields = ('title', 'nav','body','isprevious')
 def Poem(request):
     username = request.COOKIES.get('username')
     if username == '':#检测是否已经登录
@@ -88,17 +95,46 @@ def EditPoem(request):
     username = request.COOKIES.get('username')
     if username != 'god':#检测是否已经登录
         return HttpResponseRedirect('/blog/index')
-    num=-1
-    try:
-        num=request.GET['id']
-        num=int(num)
-    except:
+    if request.method == "GET":
         num=-1
-    if num>=0:
-        BlogPoem.objects.filter(id__exact=num).delete()
-    posts=BlogPoem.objects.all().order_by('-id')
-    t=render_to_response('EditPoem.html',{'posts':posts,'user':'王炳宁'},context_instance=RequestContext(request))
-    return t
+        try:
+            num=request.GET['id']
+            num=int(num)
+        except:
+            num=-1
+        try:
+            alter=request.GET['aid']
+            alter=int(alter)
+        except:
+            alter=-1
+        if num>=0:
+            BlogPoem.objects.filter(id__exact=num).delete()
+        if alter>=0:
+            post=BlogPoem.objects.get(id=alter)#要改变的
+            global preeditid
+            preeditid=post.id
+            uf = PoemFormModel(instance=post)
+            return render_to_response('PoemPost.html',{'uf':uf})
+        posts=BlogPoem.objects.all().order_by('-id')
+        t=render_to_response('EditPoem.html',{'posts':posts,'user':'王炳宁'},context_instance=RequestContext(request))
+        return t
+    else:
+        uf = PoemFormModel(request.POST)
+        if uf.is_valid():
+            global preeditid
+            #获取表单信息
+            title = uf.cleaned_data['title']
+            nav = uf.cleaned_data['nav']
+            body = uf.cleaned_data['body']
+            isprevious=uf.cleaned_data['isprevious']
+            if isprevious==u'True':
+                isprevious=True
+            else:
+                isprevious=False
+            BlogPoem.objects.filter(id=preeditid).update(title=title,nav=nav,body=body,isprevious=isprevious)
+            return HttpResponseRedirect('/blog/shuo')
+        else:
+            return HttpResponseRedirect('/blog/index')
 def __getdiv(instr):
     regx='。|！'
     afters='。；'
